Add `_checks.py`
